+++ src/components/user-routers-views/index.ts
@@ -1,16 +1,18 @@
 import { Options } from 'vue-class-component';
-import template from "./sign-in.vue";
+import template from "./user-routers-views.vue";
 import baseVue from '@/utils/base-vue';
+import RouterViews from '@/components/router-views';
 @Options({
     mixins: [template],
-    name: 'SignIn',
+    name: 'UserRouters',
+    components: {
+        RouterViews
+    }
 })
-export default class SignIn extends baseVue {
-    public setTime: any;
+export default class UserRouters extends baseVue {
     public isShowPageName: boolean =false;
 
     public mounted():void{
         this.isShowPageName = true;
     }
-
 }
