+++ src/layouts/user/sign-in/send-reset-email/index.ts
@@ -14,6 +14,7 @@ import { Field, Form, useField } from 'vee-validate';
 
 })
 export default class SendResetEmail extends baseVue {
+    public isLoding:boolean = false;
     public  returnToLogin():void{
         this.setRouter({ name: 'ChangeUserSignIn'})
     }
