+++ src/layouts/user/sign-up/resend-email/index.ts
