+++ src/layouts/user/sign-up/resend-email/resend-email.vue
