+++ src/layouts/user/sign-up/verify-email-tips/index.ts
