+++ src/layouts/user/sign-up/verify-email-tips/verify-email-tips.vue
